strip indications of repeated scans (R) from protocol names; notify slack when deleting duplicates
