+++ heudiconv_app/assets/runner-template.sh
@@ -15,15 +15,19 @@ if [ -z "${CONTAINER_IMAGE}" ]; then
 fi
 
 # Unzip dicoms locally 
+#shellcheck disable=SC2086
 LOCAL_DICOM=$(basename ${FILES})
 # remove zip suffix
+#shellcheck disable=SC2086
 LOCAL_DICOM=${LOCAL_DICOM%.*}
+#shellcheck disable=SC2086
 unzip ${FILES} -d ${LOCAL_DICOM}
 
 # UM occasionally sends duplicated DICOM files, which will break dcmstack.
 # A known pattern is that these files are nested inside the folder of the scan that is duplicated 
 # (hence -mindepth 2), and the directory of files starts with the letter s.
 if [[ ${SITE} == UM ]]; then
+#shellcheck disable=SC2086
   duplicate=$(find ${LOCAL_DICOM} -mindepth 2 -type d)
   duplicate_dir=$(basename "${duplicate}")
   if [[ ${duplicate_dir:0:1} == s ]]; then
@@ -34,6 +38,7 @@ fi
 
 if [[ ${SITE} == UC ]]; then
 
+  #shellcheck disable=SC2086
   echo singularity exec \
     --cleanenv \
     --env PREPEND_PATH=/opt/dcm2niix-UC/bin \
@@ -49,6 +54,7 @@ if [[ ${SITE} == UC ]]; then
     ${SESSION_FOR_LONGITUDINAL} ${BIDS} ${OVERWRITE} \
     ${DATALAD} ${DCMCONFIG}
 
+  #shellcheck disable=SC2086
   singularity exec \
     --cleanenv \
     --env PREPEND_PATH=/opt/dcm2niix-UC/bin \
@@ -66,6 +72,7 @@ if [[ ${SITE} == UC ]]; then
 
 else
 
+  #shellcheck disable=SC2086
   echo singularity exec \
     --cleanenv \
     -B "${BIND_DIR}":"${BIND_DIR}" \
@@ -80,6 +87,7 @@ else
     ${SESSION_FOR_LONGITUDINAL} ${BIDS} ${OVERWRITE} \
     ${DATALAD} ${DCMCONFIG}
 
+  #shellcheck disable=SC2086
   singularity exec \
     --cleanenv \
     -B "${BIND_DIR}":"${BIND_DIR}" \
@@ -97,6 +105,7 @@ else
 fi
 
 # Remove local dicom directory
+#shellcheck disable=SC2086
 rm -rf ${LOCAL_DICOM}
 
 # add bval, bvec, betc to .bidsignore
@@ -139,18 +148,29 @@ readonly FILE_EDITS=("${OUTDIR}"/sub-*/ses-*/*/*) \
 
 # Delete duplicate scans if flag is set
 echo "delete duplicates flag set to: ${DELETE_DUPLICATES}"
-if [ ${DELETE_DUPLICATES} == 1 ]; then
-  # delete duplicte scans
-  echo "removing duplicate scans"
-  rm -rf ${OUTDIR}/sub-*/ses-*/*/*_dup*
-  # remove duplicate scans from scans.tsv
-  sed -i '/_dup/d' ${OUTDIR}/sub-*/ses-*/*scans.tsv
-else
-  echo "adding duplicate scans to bids ignore"
-  # otherwise add to bids ignore
-  echo "${OUTDIR}/sub-*/ses-*/*/*_dup*" >> .bidsignore
-  # remove duplicate scans from scans.tsv
-  sed -i '/_dup/d' ${OUTDIR}/sub-*/ses-*/*scans.tsv
+mapfile -t dups <<< "$(find "${OUTDIR}" -type f -name "*dup*")"
+if [[ ${#dups[@]} -gt 0 ]]; then
+  # post about found duplicates to slack channel
+  msg="duplicate scans found: ${dups[*]}"
+  #shellcheck disable=SC2086
+  singularity exec \
+    -B "${BIND_DIR}":"${BIND_DIR}" \
+    --cleanenv docker://${CONTAINER_IMAGE} python3 log.py "${msg}" ${POST}
+  if [[ ${DELETE_DUPLICATES} == 1 ]]; then
+    # delete duplicte scans
+    echo "removing duplicate scans" 
+    rm -rf "${dups[@]}"
+    # remove duplicate scans from scans.tsv
+    sed -i '/_dup/d' "${OUTDIR}"/sub-*/ses-*/*scans.tsv
+  else
+    echo "adding duplicate scans to bids ignore"
+    # otherwise add to bids ignore
+    echo "${OUTDIR}/sub-*/ses-*/*/*_dup*" >> .bidsignore
+    # remove duplicate scans from scans.tsv
+    sed -i '/_dup/d' "${OUTDIR}"/sub-*/ses-*/*scans.tsv
+  fi
+  else
+    echo "no duplicate scans found"
 fi
 
 if [[ "${PHANTOM}" == "--no-phantom" ]]; then
@@ -211,6 +231,7 @@ if [[ ${CHECK_JSONS} == 1 ]]; then
   # the check is a bit messy. Previously, $SITE could reliably distinguish acquisition protocol. Now, sites
   # have both a patient protocol and a phantom protocol, which always differ. So, the checks must
   # be divided by whether we're dealing with a phantom scan or not.
+  #shellcheck disable=SC2086
   singularity exec \
     --cleanenv \
     -B "${BIND_DIR}":"${BIND_DIR}" \
